Beam Completion Plugin: Project Setup
