+++ plugins/beam-code-completion-plugin/.run/Run IDE with Plugin.run.xml
