+++ plugins/beam-code-completion-plugin/settings.gradle.kts
@@ -4,24 +4,16 @@
  * distributed with this work for additional information
  * regarding copyright ownership.  The ASF licenses this file
  * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
+ * License); you may not use this file except in compliance
  * with the License.  You may obtain a copy of the License at
  *
  *     http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
+ * distributed under the License is distributed on an AS IS BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
 
-
-import com.intellij.testFramework.fixtures.BasePlatformTestCase;
-import org.junit.Assert;
-
-public class JavaSDKCompletionContributorTestcase extends BasePlatformTestCase {
-    public void testAutoPopupCompletions() {
-        Assert.assertEquals(2,2);
-    }
-}
+rootProject.name = "beam-code-completion-plugin"
