+++ plugins/beam-code-completion-plugin/src/main/java/BeamJavaSDKPattern.java
@@ -4,16 +4,31 @@
  * distributed with this work for additional information
  * regarding copyright ownership.  The ASF licenses this file
  * to you under the Apache License, Version 2.0 (the
- * License); you may not use this file except in compliance
+ * "License"); you may not use this file except in compliance
  * with the License.  You may obtain a copy of the License at
  *
  *     http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an AS IS BASIS,
+ * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
 
-rootProject.name = "java-sdk-code-completion-plugin"
+import com.intellij.patterns.PatternCondition;
+import com.intellij.psi.PsiElement;
+import com.intellij.util.ProcessingContext;
+import org.jetbrains.annotations.NotNull;
+
+
+public class BeamJavaSDKPattern extends PatternCondition<PsiElement> {
+    BeamJavaSDKPattern() {
+        super("javaSDKPattern()");
+    }
+
+    @Override
+    public boolean accepts(@NotNull PsiElement psiElement, ProcessingContext context) {
+        return false;
+    }
+}
